revert(aiguard): remove Anthropic auto-instrumentation
Anthropic raw-response lifecycle handling still has unresolved performance and transport-specific correctness trade-offs under parse/asResponse concurrency. Revert the integration before the v5 release rather than ship incomplete output blocking.

Refs: https://github.com/DataDog/dd-trace-js/pull/9219
