+++ .github/workflows/instrumentation.yml
@@ -61,16 +61,6 @@ jobs:
       - uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7.0.0
       - uses: ./.github/actions/instrumentations/test
 
-  instrumentation-anthropic-lifecycle:
-    runs-on: ubuntu-latest
-    permissions:
-      id-token: write
-    env:
-      PLUGINS: anthropic-lifecycle
-    steps:
-      - uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7.0.0
-      - uses: ./.github/actions/instrumentations/test
-
   instrumentation-aws-sdk:
     runs-on: ubuntu-latest
     permissions:
