+++ packages/dd-trace/src/aiguard/integrations/index.js
@@ -1,6 +1,5 @@
 'use strict'
 
-const anthropic = require('./anthropic')
 const openai = require('./openai')
 const vercelAi = require('./vercel-ai')
 
@@ -15,7 +14,6 @@ let isEnabled = false
 function enable (aiguard, block) {
   if (isEnabled) return
 
-  anthropic.enable(aiguard, block)
   openai.enable(aiguard, block)
   vercelAi.enable(aiguard, block)
 
@@ -25,14 +23,12 @@ function enable (aiguard, block) {
 function disable () {
   vercelAi.disable()
   openai.disable()
-  anthropic.disable()
   isEnabled = false
 }
 
 module.exports = {
   enable,
   disable,
-  anthropic,
   openai,
   vercelAi,
 }
